Fix unreachable loop error
